refactor: adjust and fix worker related tests
