--- src/dal/index.ts
@@ -1,2 +1,0 @@
-export * as pair from './pair';
-export * as token from './token';
