+++ src/database/pair/pair-db.service.ts
@@ -1,11 +1,26 @@
 import { Injectable } from '@nestjs/common';
 import { Pair, Token } from '@prisma/client';
 import { PrismaService } from '../prisma.service';
-import { validTokenCondition } from '../../dal/token';
 import { ContractAddress } from '../../lib/utils';
+import { validTokenCondition } from '../token/token-db.service';
 
 export type PairWithTokens = { token0: Token; token1: Token } & Pair;
 export type CountMode = 'all' | 'listed' | 'synchronized';
+
+const tokenCondition = (showInvalidTokens: boolean, onlyListed?: boolean) => ({
+  is: {
+    ...(onlyListed ? { listed: true } : {}),
+    ...(showInvalidTokens ? {} : validTokenCondition),
+  },
+});
+
+const tokensCondition = (showInvalidTokens: boolean, onlyListed?: boolean) => {
+  const condition = tokenCondition(showInvalidTokens, !!onlyListed);
+  return {
+    token0: condition,
+    token1: condition,
+  };
+};
 
 @Injectable()
 export class PairDbService {
@@ -22,7 +37,7 @@ export class PairDbService {
 
   getAllWithCondition(showInvalidTokens: boolean, onlyListed?: boolean) {
     return this.prisma.pair.findMany({
-      where: this.tokensCondition(showInvalidTokens, onlyListed),
+      where: tokensCondition(showInvalidTokens, onlyListed),
       include: {
         token0: true,
         token1: true,
@@ -50,7 +65,7 @@ export class PairDbService {
 
   getAllWithLiquidityInfo(showInvalidTokens: boolean, onlyListed?: boolean) {
     return this.prisma.pair.findMany({
-      where: this.tokensCondition(showInvalidTokens, onlyListed),
+      where: tokensCondition(showInvalidTokens, onlyListed),
       include: {
         token0: true,
         token1: true,
@@ -75,7 +90,7 @@ export class PairDbService {
   count(showInvalidTokens: boolean, mode?: CountMode) {
     return this.prisma.pair.count({
       where: {
-        ...this.tokensCondition(showInvalidTokens, mode === 'listed'),
+        ...tokensCondition(showInvalidTokens, mode === 'listed'),
         ...(mode === 'synchronized' ? { synchronized: true } : {}),
       },
     });
@@ -152,25 +167,4 @@ export class PairDbService {
       data: { synchronized: false },
     });
   }
-
-  private tokenCondition = (
-    showInvalidTokens: boolean,
-    onlyListed?: boolean,
-  ) => ({
-    is: {
-      ...(onlyListed ? { listed: true } : {}),
-      ...(showInvalidTokens ? {} : validTokenCondition),
-    },
-  });
-
-  private tokensCondition = (
-    showInvalidTokens: boolean,
-    onlyListed?: boolean,
-  ) => {
-    const condition = this.tokenCondition(showInvalidTokens, !!onlyListed);
-    return {
-      token0: condition,
-      token1: condition,
-    };
-  };
 }
