+++ test/contracts.spec.ts
@@ -1,5 +1,5 @@
 import { getContext, Context } from '../src/lib/contracts';
-import { mockupEnvVars, TEST_NET_VARS } from './utils/env.mockups';
+import { mockupEnvVars, TEST_NET_VARS } from './utils';
 
 describe('with real Context on testnet', () => {
   let context: Context | null = null;
@@ -20,11 +20,13 @@ describe('with real Context on testnet', () => {
     const { decodedResult: factoryAddress } = await ctx().router.factory();
     expect(factoryAddress).toBe(process.env.FACTORY_ADDRESS);
   });
+
   it('should have at least 16 pairs', async () => {
     expect(
       (await ctx().factory.get_all_pairs()).decodedResult.length,
     ).toBeGreaterThanOrEqual(16);
   });
+
   it('pair should return right token addresses', async () => {
     const { decodedResult: allPairs } = await ctx().factory.get_all_pairs();
     const pairAddress = allPairs[allPairs.length - 1];
@@ -42,6 +44,7 @@ describe('with real Context on testnet', () => {
       'ct_b7FZHQzBcAW4r43ECWpV3qQJMQJp5BxkZUGNKrqqLyjVRN3SC',
     );
   });
+
   it('regular aex9 token should have right metaInfo', async () => {
     const tokenMethods = await ctx().getToken(
       'ct_7tTzPfvv3Vx8pCEcuk1kmgtn4sFsYCQDzLi1LvFs8T5PJqgsC',
@@ -54,6 +57,7 @@ describe('with real Context on testnet', () => {
       symbol: 'TAEX9-B',
     });
   });
+
   it('WAE token should have right metaInfo', async () => {
     const tokenMethods = await ctx().getToken(
       'ct_JDp175ruWd7mQggeHewSLS1PFXt9AzThCDaFedxon8mF8xTRF',
