+++ test/utils/db.ts
@@ -1,6 +1,7 @@
-import db from '../../src/dal/client';
-export const clean = async () => {
-  await db.pairLiquidityInfo.deleteMany();
-  await db.pair.deleteMany();
-  await db.token.deleteMany();
+import { PrismaService } from '../../src/database/prisma.service';
+
+export const clean = async (prismaService: PrismaService) => {
+  await prismaService.pairLiquidityInfo.deleteMany();
+  await prismaService.pair.deleteMany();
+  await prismaService.token.deleteMany();
 };
