+++ test/utils/index.ts
@@ -1,13 +1,14 @@
+import { PrismaService } from '../../src/database/prisma.service';
+
 export * from './context.mockup';
 export * from './env.mockups';
 export * from './db';
-import db from '../../src/dal/client';
 
 export const sortByAddress = (xs: { address: string }[]) =>
   [...xs].sort((a, b) => a.address.localeCompare(b.address));
 
-export const listToken = (address: string) =>
-  db.token.update({
+export const listToken = (prismaService: PrismaService, address: string) =>
+  prismaService.token.update({
     where: {
       address,
     },
